Enhance `Authorize` and trade execution flow
- Update `Authorize` to return an `Account` struct with `ID` and `Currency`.
- Refactor `Buy` and `Sell` methods to use `Position` struct for cleaner parameter management.
- Introduce `Account` struct in executor package for holding account details.
- Update `ExecuteStrategy` to utilize account currency and refactor position handling.

+++ pkg/core/executor/strategy.go
@@ -15,7 +15,7 @@ type Strategy struct {
 	Symbol       string
 	Amount       float64
 	Type         StrategyType
-	Leverage     int
+	Leverage     float64
 	CheckToOpen  func(tick signal.Tick) bool
 	CheckToClose func(tick signal.Tick) bool
 }
